Added a dummy implementation for SSL_stacktrace_get_all_frame_names, and did some minor tweaks here and there
svn path=/spoofax/trunk/spoofax/org.spoofax.interpreter.core/; revision=18227

+++ org.spoofax.interpreter.core/src/java/org/spoofax/interpreter/library/ssl/SSL_real.java
@@ -34,7 +34,7 @@ public class SSL_real extends AbstractPrimitive {
         if(Tools.isTermInt(tvars[0]))
         {
             IStrategoInt a = (IStrategoInt) tvars[0];
-            env.setCurrent(env.getFactory().makeReal((double) a.intValue()));
+            env.setCurrent(env.getFactory().makeReal(a.intValue()));
             return true;
         }
